Alter errors to occur on methods in order
Previously, I waited till we parsed all methods so that I knew the final
location a method should be defined in so the error message could be
helpful in finding it's final location.

However, with the following example, we would show an error saying that
method `apple` should show before method `bear`, when it was already
defined in that order.

```
def cat
end

def apple
end

def bear
end
```

I changed the errors to be shown on nodes in method definition order
instead of final order. So now we would show errors for the methods
`apple` and `bear` saying they should each show before the method `cat`.

Since each node that needed to move for autocorrect was no longer
being marked as an error, I updated the returned node info for autocorrection
to include the method that triggered the issue as well as the location
it is trying to be moved to so that both methods are eventually moved.

This created some duplicate operations so I cache which corrections have
been made so I only perform each once.