Add test and account for missed method order combination
At this point, the `replacements` code should be looked at again
since it is returning a different object than is needed to handle the
actual replacements.

The issue is there is a difference between the nodes we are listing as
invalid and the nodes that we want to autocorrect in regards to their
final order in the file.